transcribe/requirements.txt
google-cloud-speech==2.20.0
protobuf==3.20.3
firebase-admin==6.1.0